requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.3
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
colorama==0.4.1
curses==2.2
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
future==0.17.1
gast==0.2.2
grpcio==1.25.0
gym==0.12.5
h5py==2.9.0
imutils==0.5.3
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.21
jedi==0.13.3
Jinja2==2.10.1
joblib==0.14.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
nbconvert==5.5.0
nbformat==4.4.0
nltk==3.4.1
notebook==5.7.8
npm==0.1.1
numpy==1.17.3
opencv-python==4.1.1.26
optional-django==0.1.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
pyglet==1.3.2
Pygments==2.4.2
pylint==2.4.4
pynput==1.4.5
pyparsing==2.4.2
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.5.1
scikit-learn==0.21.3
scipy==1.1.0
seaborn==0.10.0
Send2Trash==1.5.0
six==1.13.0
soupsieve==1.9.5
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tflearn==0.3.2
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
typed-ast==1.4.1
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.2
xlrd==1.2.0
